web/requirements.txt
Django==4.2
djangorestframework
mysqlclient
python-decouple>=3.0